Added client-side encryption
(This is exactly the previous commit, but with a description in English.)
Now the data can be encrypted at the software level before transmission over the network, that is, without SSL.
Encryption is performed using the 'cryptography' package: https://cryptography.io/ 
Fernet algorithm is used. (symmetric encryption)
It is assumed that the key is pre-generated and known to all clients involved in the encryption (file "key").